+++ src/TRegx/CleanRegex/Match/Details/Match.php
@@ -1,0 +1,9 @@
+<?php
+namespace TRegx\CleanRegex\Match\Details;
+
+/**
+ * @deprecated Use {@see \TRegx\CleanRegex\Match\Details\Detail} instead
+ */
+interface Match extends Detail
+{
+}