+++ src/TRegx/CleanRegex/Match/Details/ReplaceMatch.php
@@ -1,0 +1,9 @@
+<?php
+namespace TRegx\CleanRegex\Match\Details;
+
+/**
+ * @deprecated Use {@see \TRegx\CleanRegex\Match\Details\ReplaceDetail} instead
+ */
+interface ReplaceMatch extends ReplaceDetail, Match
+{
+}